+++ CHANGES.md
@@ -21,7 +21,7 @@
 * Helper fn to build routes from given route specs
 * Allow non-literal string URI-patterns in `calfpath.core/->uri`
 * Fix `calfpath.route/update-fallback-400` to add fallback 400 route on one or more URI entry, instead of all
-* [TODO] BREAKING CHANGE: Drop `calfpath.core/make-uri-handler` in favor of Calfpath routes
+* BREAKING CHANGE: Drop `calfpath.core/make-uri-handler` in favor of Calfpath routes
 
 
 ## 0.3.0 / 2016-May-27